+++ crates/reading/src/lib.rs
@@ -47,4 +47,7 @@ pub use differential_dataflow::operators::iterate::Variable as RecVariable;
 pub mod epoch;
 pub use epoch::Epoch;
 pub type Time = Epoch;
-pub type Iter = u16;
+/// Inner (fixpoint) iteration counter. u32: a divergent recursion used to
+/// WRAP the old u16 at 65535, silently corrupting nested timestamps (the
+/// Product lattice order is violated by wraparound).
+pub type Iter = u32;
